backend/portfolio_agents/requirements.txt
# CrewAI Multi-Agent Portfolio Construction System Requirements

# Core CrewAI Framework
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-anthropic==0.1.15

# Anthropic AI
anthropic==0.25.9

# Data Processing & Analysis
pandas==2.2.2
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0

# Financial Data APIs
yfinance==0.2.18
alpha-vantage==2.3.1
finnhub-python==2.4.20
sec-edgar-api==2.1.0

# Technical Analysis
ta-lib==0.4.26
pandas-ta==0.3.14b0

# Sentiment Analysis & NLP
transformers==4.40.2
torch==2.0.1
finbert==0.1.6
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Knowledge Graph & Database
neo4j==5.19.0
py2neo==2021.2.4
rdflib==7.0.0
owlrl==6.0.2
sparqlwrapper==2.0.0
networkx==3.1

# Portfolio Optimization
cvxpy==1.4.3
pypfopt==1.5.5

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
bokeh==3.2.1

# Web Framework & APIs
fastapi==0.111.0
uvicorn==0.22.0
requests==2.31.0
httpx==0.24.1

# Configuration & Environment
pydantic==2.7.1
python-dotenv==1.0.0
pyyaml==6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0

# Logging & Monitoring
loguru==0.7.0

# Redis for caching
redis==4.5.5

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.0

# Jupyter notebooks for research
jupyter==1.0.0
ipykernel==6.23.1